requirements.txt
Flask==3.1.0
anthropic==0.43.0
python-dotenv==1.0.1
edge-tts==7.2.8
gspread==6.1.4
google-auth==2.38.0
